requirements.txt
aiohttp>=3.8.0
pandas>=1.3.0
tqdm>=4.65.0
psutil>=5.9.0
pyyaml>=6.0.0
openpyxl>=3.0.0  # 用于Excel文件支持 